Refactor IXyloxTrack
Made no sense to have this instead of null. Didn't want to introduce having to check.

So instead the interface IXyloxTrack is now a global interface for all services which simply has a message property that can be used when handling errors or simply returning a status message of sorts.